# milktea/skills/milktea-openapi.yaml
openapi: 3.1.0
info:
  title: MilkTea HTTP API — thiết kế một quán
  version: 2.0.0
  description: 'Hợp đồng thiết kế để triển khai; không xác nhận endpoint đã có code. Domain: domain_system.md; HTTP
    envelope: system_architecture.md; security: author_rule.md. Không mô tả STOMP trong OpenAPI. QR read-only; session
    mở khi order/CASHIER open; thanh toán linh hoạt; chỉ CASHIER close. x-allowed-account-roles chỉ ràng buộc nhánh
    xác thực tài khoản, KHÔNG phải điều kiện role cho guest. Guest được cho phép theo nhánh qrAccess/tableAccess/counterAccess
    và phạm vi resource tương ứng; luôn kiểm Domain. Tooling OpenAPI không tự enforce role hay ownership.

    Nguồn: báo cáo MilkTea cập nhật 05/10/2026, 41 UC/T01–T60. Hợp đồng thiết kế; chưa xác nhận các endpoint đã
    được triển khai. v2 bổ sung UC23 thông tin quán, UC30 ngưỡng theo vị trí, UC39 sự cố/tiêu hao pha bù; version
    BIGINT truyền chuỗi; giữ giỏ/context trong kết quả đặt bàn.'
servers:
- url: http://localhost:8080
  description: Ví dụ local; dùng SERVER_PORT và origin thực của project.
paths:
  /api/v1/auth/csrf:
    get:
      operationId: readCsrf
      summary: Lấy CSRF cho form/AJAX/STOMP browser
      description: Lấy CSRF cho form/AJAX/STOMP browser
      tags:
      - Auth
      security: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/CsrfView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      x-use-cases:
      - UC01
      - UC04
  /api/v1/auth/register:
    post:
      operationId: registerCustomer
      summary: Đăng ký CUSTOMER
      description: Không nhận role. Không tự tạo account nhân viên/admin.
      tags:
      - Auth
      security: []
      responses:
        '201':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/AccountView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      x-use-cases:
      - UC02
  /api/v1/auth/login:
    post:
      operationId: loginAccount
      summary: Đăng nhập và đặt cookie JWT HttpOnly
      description: Response data không chứa raw JWT; đặt Set-Cookie milktea_access_token. Lấy CSRF mới sau login.
        Postman có thể sử dụng cookie nhận được.
      tags:
      - Auth
      security: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/AccountView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      x-use-cases:
      - UC01
  /api/v1/auth/logout:
    post:
      operationId: logoutAccount
      summary: Đăng xuất, tăng tokenVersion và đóng socket account
      description: Revoke account JWT, xóa cookie; không kết thúc TableSession.
      tags:
      - Auth
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/MessageView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      x-use-cases:
      - UC04
...
  /api/v1/auth/forgot-password:
    post:
      operationId: requestPasswordReset
      summary: Yêu cầu reset bằng mail
      description: Thông báo cùng dạng dù email không tồn tại; token chỉ gửi mail.
      tags:
      - Auth
      security: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/MessageView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      x-use-cases:
      - UC03
...
components:
  securitySchemes:
    accountCookie:
      type: apiKey
      in: cookie
      name: milktea_access_token
      description: JWT account HttpOnly; role/state/version enforced server-side.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tableAccess:
      type: apiKey
      in: header
      name: X-Table-Token
      description: Opaque access scoped to one currently OPEN TableSession, not an account role.
    counterAccess:
      type: apiKey
      in: header
      name: X-Counter-Token
      description: Opaque access scoped to one counter order; client keeps only in current page RAM.
    qrAccess:
      type: apiKey
      in: header
      name: X-QR-Code
      description: Opaque QR code identifies a table; never a staff credential.
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      example: '5'
...
    Role:
      type: string
      enum:
      - ADMIN
      - CASHIER
      - KITCHEN
      - CUSTOMER
...
    ErrorResponse:
      type: object
      properties:
        success:
          const: false
        error:
          type: object
          properties:
            code:
              type: string
              maxLength: 80
            message:
              type: string
              maxLength: 1000
            fieldErrors:
              type: object
              additionalProperties:
                type: string
                maxLength: 1000
          additionalProperties: false
          required:
          - code
          - message
          - fieldErrors
        traceId:
          type: string
          maxLength: 100
      additionalProperties: false
      required:
      - success
      - error
      - traceId
    CsrfView:
      type: object
      properties:
        headerName:
          const: X-CSRF-TOKEN
        token:
          type: string
          maxLength: 1024
      additionalProperties: false
      required:
      - headerName
      - token
    MessageView:
      type: object
      properties:
        message:
          type: string
          maxLength: 1000
      additionalProperties: false
      required:
      - message
    AccountView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
          maxLength: 120
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        active:
          type: boolean
      additionalProperties: false
      required:
      - id
      - name
      - email
      - role
      - active
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          maxLength: 128
      additionalProperties: false
      required:
      - email
      - password
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
      additionalProperties: false
      required:
      - name
      - email
      - password
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      additionalProperties: false
      required:
      - email
...
  responses:
    Error400:
      description: HTTP 400; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: HTTP 401; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: HTTP 403; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: HTTP 404; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: HTTP 409; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Error500:
      description: HTTP 500; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- accountCookie: []
- bearerAuth: []
...